silo: inline xb_silo_get_node() across modules
The xb_silo_get_node() function was marked for inline but was an exported
module. This limits what will actually get inlined to the current module
and LTO doesn't seem to be doing much beyond that.

However, we can store the offset to the direct position from the self
struct to the data array and load that without roundtripping through the
private structure (nor having to expose that).

This allows xb-silo-query.c to inline the get_node() calls as well where
they are in the hot path.